Order Completed

+++ src/main/java/com/example/demo/services/libreria/BookServiceImpl.java
@@ -44,6 +44,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deletedById(Integer id) {
+        if (!bookRepository.existsById(id)) {
+            throw new RuntimeException("Book not founded, with id - " + id);
+        }
         bookRepository.deleteById(id);
     }
